chore: move a1z cli
