+++ dimos/cli/dimos.py
@@ -49,6 +49,7 @@ import requests
 import typer
 
 from dimos.agents.mcp.mcp_adapter import McpAdapter, McpError
+from dimos.cli.a1z import app as a1z_app
 from dimos.cli.cache import app as cache_app
 from dimos.cli.shell import shell
 from dimos.constants import CONFIG_DIR, LOG_DIR
@@ -60,7 +61,6 @@ from dimos.mapping.cli.pose_fill import main as _map_pose_fill_main
 from dimos.mapping.cli.rename import main as _map_rename_main
 from dimos.mapping.cli.replay import main as _map_replay_main
 from dimos.mapping.cli.replay_marker import main as _map_replay_marker_main
-from dimos.robot.manipulators.a1z.cli import app as a1z_app
 from dimos.robot.manipulators.piper.cli import app as piper_app
 from dimos.robot.unitree.go2.cli.go2tool import app as go2tool_app
 from dimos.utils.cache import cache_usage_locked
